Prevent watchman version check from throwing
watchman releases are no longer tagged with a semver valid version
number. This is causing ember-cli to throw a TypeError when attempting
to check if minimum version requirements are satisfied.

this change attempts to check if the installed version of watchman is
semver valid prior to seeing if they satisfy requirements. if the
version is invalid, default to as satisying requirements.

- ref https://github.com/facebook/watchman/releases